Mohammed Salah 09-04-2022 03:57 AM

+++ database/migrations/2022_04_02_000716_create_courses_table.php
@@ -22,7 +22,7 @@ class CreateCoursesTable extends Migration
             $table->longText('teacher_ar');
             $table->longText('teacher_en')->nullable();
             $table->integer('section_count');
-            $table->integer('section_time');
+            $table->decimal('section_time',10,2);
             $table->date('course_date');
             $table->tinyInteger('status')->comment('2 => Avtive || 1 => Stopped');
             $table->longText('main_image')->nullable();
